bool to int

+++ code_prediction/funfam.py
@@ -278,7 +278,7 @@ class FunFam:
         for i,member in enumerate(self.members):
             if not member.binding_annotation:
                 continue
-            annotation = self.map_from_alignment_to_sequence(member.aligned_sequence, self.binding_sites[member.id])
+            annotation = self.map_from_alignment_to_sequence(member.aligned_sequence, self.binding_sites[member.id]).astype(int)
             cum_scores = self.map_score_to_sequence(member.aligned_sequence, self.predictions_cum_scores[member.id])
             clust_scores = self.map_score_to_sequence(member.aligned_sequence, self.predictions_cluster_coeff[member.id])
             if sum(annotation) == 0:
